+++ core-ng-search-test/src/main/java/core/framework/search/module/TestSearchConfig.java
@@ -11,14 +11,17 @@ import org.apache.logging.log4j.core.config.ConfigurationFactory;
  * @author neo
  */
 public class TestSearchConfig extends SearchConfig {
+    // only start one local node for testing to reduce resource overhead,
+    // only breaking case is that multiple search() using same index name, then if one unit test operates both ElasticSearchType will result in conflict or merged results
+    // this can be avoided by designing test differently
+    private static HttpHost localESHost;
+
     static {
         // refer to io.netty.util.NettyRuntime.AvailableProcessorsHolder.setAvailableProcessors
         // refer to org.elasticsearch.transport.netty4.Netty4Utils.setAvailableProcessors
         // netty only allows set available processors once
         System.setProperty("es.set.netty.runtime.available.processors", "false");
     }
-
-    private HttpHost host;
 
     @Override
     protected void initialize(ModuleContext context, String name) {
@@ -27,14 +30,19 @@ public class TestSearchConfig extends SearchConfig {
     }
 
     private void startLocalElasticSearch(ModuleContext context) {
-        var server = new LocalElasticSearch();
-        host = server.start();
-        context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> server.close());
+        synchronized (TestSearchConfig.class) {
+            // in test env, config is initialized in order and within same thread, so no threading issue
+            if (localESHost == null) {
+                var server = new LocalElasticSearch();
+                localESHost = server.start();
+                context.shutdownHook.add(ShutdownHook.STAGE_7, timeout -> server.close());
+            }
+        }
     }
 
     @Override
     public void host(String host) {
-        search.hosts = new HttpHost[]{this.host};
+        search.hosts = new HttpHost[]{localESHost};
     }
 
     // es refers to log4j core directly in org.elasticsearch.common.logging.Loggers, this is to bridge es log to coreng logger