+++ core-ng-search-test/src/test/java/core/framework/search/TestModule.java
@@ -23,5 +23,10 @@ public class TestModule extends AbstractTestModule {
         initSearch.putIndex("document", "search-test/document-index.json");
         initSearch.putIndexTemplate("document", "search-test/document-index-template.json");
         initSearch.flush("document");
+
+        // test multiple search in one app
+        search = config(SearchConfig.class, "other");
+        search.host("localhost");
+        search.type(TestDocument.class);
     }
 }